Fix some bugs

+++ queryscript/src/compile/builtin_types.rs
@@ -70,7 +70,7 @@ fn max<R>(value R) -> R = sql;
 fn count<R>(value R) -> bigint = sql;
 fn sum<R>(value R) -> SumAgg<R> = sql;
 fn avg<R>(value R) -> double = sql;
-fn grouping<R>(value R) -> tinyint = sql;
+fn grouping<R>(value R) -> bigint = sql;
 
 -- Window functions
 fn row_number<R>() -> bigint = sql;
